requirements.txt
streamlit
PyPDF2
pdfplumber
Pillow
streamlit-lottie


